Simplify user permissions with only two permissions and add specific errors translations for player view
There were permissions for viewing the block, the list and the player. As the block can be visualized only by enrolled users, there is no need to have specific permissions for accessing the block. All we need is a permission to edit the block and its content.
Also change the translation keys used by player page errors which were relative to list page instead of player page.