Change the Web Service connection error message
Using a wrong Web Service configuration was leading to a message informing
that the connection to the Web Service couln't be established. This same
error message was displayed when getting the list of custom properties
failed due to missing scopes for the client. Message is now clearer to
check both Web Service configuration and client's scopes.