Use template engine to render player page
Player page was rendered manually. As Moodle
now includes Mustache template engine, the player
page is now rendered using it.
The player is not rendered anymore, the player
page only displays a link that might be
transformed automatically into a player if Moodle
OpenVeo Player plugin is installed.
Also not that the dimension of the generated
player have been forced using CSS as a bug into
Moodle unables usage of percentage when setting
the dimension in the video URL.